pyproject.toml
[tool.poetry]
name = "TODO"
version = "0.2.0"
description = ""
authors = ["Steve Wyckoff <stevewyckoffdev@gmail.com>"]
readme = "README.md"
classifiers = [
    "Development Status :: 4 - Beta",
    "License :: OSI Approved :: MIT",
    "Natural Language :: English",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
]

packages = [
    { include = "common", from = "src" }
]

[tool.poetry.dependencies]
python = "^3.12"
coloredlogs = "^15.0.1"
python-json-logger = "^2.0.7"
requests = "^2.32.3"
pandas = "^2.2.3"
numpy = "^2.1.1"
pydantic = {extras = ["email"], version = "^2.9.2"}
sqlalchemy = "^2.0.35"
scipy = "^1.14.1"
colorama = "^0.4.6"
# websockets = "^13.1"
# asyncio = "^3.4.3"
pydantic-settings = "^2.5.2"
# faker = "^30.1.0"

[tool.poetry.dev-dependencies]
pytest = "^6.2"
ipykernel = "^6.29.5"
ipython-sql = "^0.5.0"
pdoc3 = "^0.11.1"
python-semantic-release = "^9.9.0"
python-dotenv = "^1.0.1"
black = "^24.8.0"
darker = "^2.1.1"
ipdb = "^0.13.13"
pdbpp = "^0.10.3"
debugpy = "^1.8.6"
ipython = "^8.28.0"
jupyterlab = "^4.2.5"
notebook = "^7.2.2"
matplotlib = "^3.9.2"
types-colorama = "^0.4.15.20240311"

[tool.poetry.scripts]
entry = "src.server.app:run" # TODO replace

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
